requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
